+++ clang-tidy/misc/StructInheritanceCheck.cpp
@@ -1,0 +1,31 @@
+//===--- StructInheritanceCheck.cpp - clang-tidy---------------------------===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#include "StructInheritanceCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace misc {
+
+void StructInheritanceCheck::registerMatchers(MatchFinder *Finder) {
+  Finder->addMatcher(cxxRecordDecl(isStruct(),isDerivedFrom(anything())).bind("struct"), this);
+}
+
+void StructInheritanceCheck::check(const MatchFinder::MatchResult &Result) {
+  const auto *MatchedDecl = Result.Nodes.getNodeAs<CXXRecordDecl>("struct");
+  diag(MatchedDecl->getLocation(), "structs should not inherit from structs or classes");
+}
+
+} // namespace misc
+} // namespace tidy
+} // namespace clang